I think I have actually fixed the bug.
The bug is caused when a key switch bounces and looses contact for a short
time. With my original code that caused the key press to be sent in twice
resulting in the occational 'eeeecho'. Once I looked through the
Teensy_keyboard code I discovered that they solved it by only removing
a key after a delay and aborting it if the key was pressed again. So I
copied them and ended up with a far more annoying bug. When a key switch
bounced the keyboard would stop registering key presses from that key.

You have no idea how long that as fucked with me for and how god damn
annoyed I got. The fix for this however was solved by another, more careful
look through the Teensy_keyboard code. They checked if they had already sent
the key through before adding it to the keyboard_keys array.

My problem was not that the key was not being registered. But that it was
being sent and never lifted as I added it to the array twice. So after
that extra presses of the key were added to the array, the OS ignored
it and then on key release the new key was released but the old still left.

I now know what the bug is, what caused it, and how to fix it.